refactor: utils -> view
